+++ lib/td_cache/redix/stream.ex
@@ -103,6 +103,16 @@ defmodule TdCache.Redix.Stream do
     Redix.command(["XTRIM", stream, "MAXLEN", count])
   end
 
+  def delete_events(stream, event_ids) do
+    Redix.command!(["XDEL", stream | event_ids])
+  end
+  def delete_if_empty(stream) do
+    case Redix.command!(["XLEN", stream]) do
+      0 -> Redix.command!(["DEL", stream])
+      _ -> :ok
+    end
+  end
+
   defp parse_results(events, opts) do
     case Keyword.get(opts, :transform) do
       :range ->